#873 - renamed packages
